+++ src/main/java/com/skyhorsemanpower/payment/payment/vo/PaymentListResponseVo.java
@@ -1,6 +1,7 @@
 package com.skyhorsemanpower.payment.payment.vo;
 
 import com.skyhorsemanpower.payment.common.PaymentStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
@@ -9,11 +10,11 @@ public class PaymentListResponseVo {
 
     private String paymentUuid;
     private String auctionUuid;
-    private double price;
+    private BigDecimal price;
     private PaymentStatus paymentStatus;
     private LocalDateTime paymentAt;
 
-    public PaymentListResponseVo(String paymentUuid, String auctionUuid, double price,
+    public PaymentListResponseVo(String paymentUuid, String auctionUuid, BigDecimal price,
         PaymentStatus paymentStatus, LocalDateTime paymentAt) {
         this.paymentUuid = paymentUuid;
         this.auctionUuid = auctionUuid;
